add another file and put the winner class and made required changes
